feat: CompanyFactory 생성
