--- src/main/java/rentCompany/Company.java
@@ -1,15 +1,0 @@
-package rentCompany;
-
-import java.util.List;
-
-public class Company {
-    private final CompanyCars companyCars;
-
-    public Company(CompanyCars companyCars) {
-        this.companyCars = companyCars;
-    }
-
-    public List<Car> getCars() {
-        return companyCars.getCars();
-    }
-}
