+++ src/main/java/rentCompany/model/CompanyCars.java
@@ -1,13 +1,18 @@
-package rentCompany;
+package rentCompany.model;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 public class CompanyCars {
     private final List<Car> cars;
 
-    public CompanyCars(List<Car> cars) {
-        this.cars = cars;
+    public CompanyCars() {
+        this.cars = new ArrayList<>();
+    }
+
+    public void add(Car car) {
+        cars.add(car);
     }
 
     public List<Car> getCars() {
