+++ src/main/java/rentCompany/model/CompanyFactory.java
@@ -1,0 +1,8 @@
+package rentCompany.model;
+
+public class CompanyFactory {
+
+    public static Company createCompany() {
+        return new Company();
+    }
+}
